Move TruncatingCharField to corehq.util.models
Preparation for using it in places outside of form processor.
