+++ corehq/util/models.py
@@ -247,6 +247,18 @@ class NullJsonField(jsonfield_JSONField):
         return self.get_default() if value is None else value
 
 
+class TruncatingCharField(models.CharField):
+    """
+    http://stackoverflow.com/a/3460942
+    """
+
+    def get_prep_value(self, value):
+        value = super().get_prep_value(value)
+        if value:
+            return value[:self.max_length]
+        return value
+
+
 class ForeignValue:
     """Property descriptor for Django foreign key refs with a primitive value
